requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
boto3==1.37.4
botocore==1.37.4
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-countries==7.6.1
django-ninja==1.3.0
django-storages==1.14.5
dnspython==2.7.0
email_validator==2.2.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
kombu==5.4.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyuca==1.2
redis==5.2.1
requests==2.32.3
s3transfer==0.11.3
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
